many fixes for collada 1.5 spec parsing and saving back to xml

+++ collada/controller.py
@@ -189,14 +189,14 @@ class Skin(Controller):
 
         bind_shape_mat = skinnode.find(tag('bind_shape_matrix'))
         if bind_shape_mat is None:
-            bind_shape_mat = numpy.identity(4, dtype=numpy.float32)
+            bind_shape_mat = numpy.identity(4, dtype=numpy.float64)
             bind_shape_mat.shape = (-1,)
         else:
             try:
                 values = [ float(v) for v in bind_shape_mat.text.split()]
             except ValueError:
                 raise DaeMalformedError('Corrupted bind shape matrix in skin')
-            bind_shape_mat = numpy.array( values, dtype=numpy.float32 )
+            bind_shape_mat = numpy.array( values, dtype=numpy.float64 )
 
         inputnodes = skinnode.findall('%s/%s'%(tag('joints'), tag('input')))
         if inputnodes is None or len(inputnodes) < 2:
